+++ animalheart/src/main/java/com/animalheart/animalheart/service/VeterinarioService.java
@@ -1,13 +1,14 @@
 package com.animalheart.animalheart.service;
 
-import com.animalheart.animalheart.entities.Veterinario;
 import java.util.List;
+
+import com.animalheart.animalheart.model.Veterinario;
 
 public interface VeterinarioService {
 
     List<Veterinario> obtenerTodos();
     Veterinario validarVeterinario(String nombreUsuario, String contrasenia);
     List<Veterinario> obtenerVeterinariosActivos();
-    Veterinario obtenerVeterinarioPorId(Integer id);
+    Veterinario obtenerVeterinarioPorId(Long id);
     
 }
